Necessary Scripts for the procedure
The folder contains different file for the procedure:
sql files -> Used for cleaning the FAERS database, they're a modified version coming from https://github.com/ltscomputingllc/faersdbstats. The logic is the same but use updated files and do not compute the safety signal. The data of interest is only the cleaned version of FAERS.
python files -> Used for retrieving, cleaning and associate the DRUG - Side effect and DRUG - Target databases